Implement my own implementation of SVD and test it
My implementation has much worse accuracy for the eigenvalues that are
numerically zero. This leads to a lot of error propagating through the
system and a poor U matrix, which means USV^T differs a lot from the
original matrix. GSL's implementation is going to be better going
forward, as it is far more accurate, but GSL only provides
implementations of
[thin
SVD](https://en.wikipedia.org/wiki/Singular_value_decomposition#Thin_SVD),
which means I will need to be careful when using it for PCA.