Rewrite the PCA implementation from scratch
This allows the eigenvectors to be extracted regardless of whether the
data is stored as row vectors or column vectors. If you have N vectors
of length D < N and want to extract the D eigenvectors of length D,
you can encode it as a row vector (X will be NxD). If you have N
vectors of length D > N and want to extract the N eigenvectors of
length D, you can encode it as a column vector (X will be DxN).